+++ src/main/java/project/volunteer/domain/recruitment/application/RecruitmentFacade.java
@@ -6,7 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import project.volunteer.domain.confirmation.application.ConfirmationService;
 import project.volunteer.domain.image.application.ImageService;
 import project.volunteer.domain.notice.application.NoticeService;
-import project.volunteer.domain.recruitmentParticipation.application.RecruitmentParticipationService;
+import project.volunteer.domain.recruitmentParticipation.application.RecruitmentParticipationUseCase;
 import project.volunteer.domain.recruitment.api.dto.request.RecruitmentRequest;
 import project.volunteer.domain.scheduleParticipation.service.ScheduleParticipationService;
 import project.volunteer.domain.sehedule.application.ScheduleCommandUseCase;
@@ -25,7 +25,7 @@ public class RecruitmentFacade {
     private final RecruitmentCommandUseCase recruitmentCommandUseCase;
     private final RecruitmentQueryUseCase recruitmentQueryUseCase;
     private final ScheduleCommandUseCase scheduleCommandUseCase;
-    private final RecruitmentParticipationService recruitmentParticipationService;
+    private final RecruitmentParticipationUseCase recruitmentParticipationUseCase;
 
 
 
@@ -46,7 +46,7 @@ public class RecruitmentFacade {
 
         scheduleCommandUseCase.deleteAllSchedule(recruitmentNo);
 
-        recruitmentParticipationService.deleteRecruitmentParticipations(recruitmentNo);
+        recruitmentParticipationUseCase.deleteRecruitmentParticipations(recruitmentNo);
